Tests:: update

+++ test.sh
@@ -2,6 +2,7 @@
 export id=()
 index=0
 pid=()
+returnSucessProcess=0
 
 pidOfScript="$(echo ${$})"
 
@@ -14,12 +15,18 @@ function trap_with_arg() {
 function func_trap() {
     echo Trapped: $1
     index=0
-    while test ${index} != ${numOfClient}
-        do
-        kill -n 30 ${pid[$index]}
-        echo "kill -n 30 ${pid[$index]}"
-        index=$(($index + 1))
-    done
+    if [ $1 = SIGUSR1 ]
+        then
+            while test ${index} != ${numOfClient}
+            do
+                kill -n 30 ${pid[$index]}
+                echo "kill -n 30 ${pid[$index]}"
+                index=$(($index + 1))
+            done
+    elif [ $1 = SIGUSR2 ]
+        then
+            returnSucessProcess=$((returnSucessProcess + 1))
+    fi
 }
 function in_id {
   for e in ${id[*]}
@@ -34,6 +41,7 @@ function in_id {
 }
 
 trap_with_arg func_trap SIGUSR1
+trap_with_arg func_trap SIGUSR2
 
 echo "pidOfScript : ${pidOfScript}"
 
@@ -54,7 +62,7 @@ numOfClient=${#id[@]}
 
 while test ${index} != ${numOfClient}
     do
-    ./nibbler -t --id ${id[$index]} --fileInput ${1} --pidTestProcess ${pidOfScript} >> /tmp/log_test.out & pid[$index]=$! > /tmp/log_pid.out
+    ./nibbler -t --id ${id[$index]} --fileInput ${1} --pidTestProcess ${pidOfScript} &> $2 & pid[$index]=$! > /tmp/log_pid.out
     echo ${id[$index]}
     index=$(($index + 1))
 done
@@ -76,12 +84,30 @@ while test ${index} != ${numOfClient}
     index=$(($index + 1))
 done
 
-while read line ; do
-    if [ $line = "kill" ]
-        then
-            for p in ${pid[*]}; do
-                kill $p > /dev/null;
-            done
-        exit
-    fi
+#while read line ; do
+#    if [ $line = "kill" ]
+#        then
+#            for p in ${pid[*]}; do
+#                kill $p > /dev/null;
+#            done
+#        exit
+#    fi
+#    returnSucessProcess
+#done
+
+
+for p in ${pid[*]}; do
+    while kill -0 ${p} > /dev/null ; do
+        sleep 0.1
+    done
 done
+
+sleep 1
+
+if [ $returnSucessProcess = $numOfClient ]
+    then
+        echo "EXIT_SUCESS $returnSucessProcess $numOfClient"
+        exit 0
+fi
+echo "EXIT_FAILURE $returnSucessProcess $numOfClient"
+exit 1
